Restructured tests to allow better expansion
